move math models out of subdirectory to avoid changing setup.py
