+++ gpkit/math_models.py
